Refactor test suite: clarify file responsibilities, add hivsim.Sim tests (#365)
- test_sim.py: rewrite as constructor/routing tests + hivsim.Sim coverage (#349)
- test_diseases.py → test_stis.py: non-HIV STI scientific validation
- test_hiv_natural_history_verification.py → test_hiv.py: HIV scientific validation
- test_networks.py: add test_msm_network
- test_examples.py: merge into test_sim.py, delete
- Add/update docstrings clarifying scope for all test files
