--- tests/test_examples.py
@@ -1,55 +1,0 @@
-"""
-Tests for hivsim_examples package and hivsim.demo()
-"""
-import numpy as np
-import hivsim
-
-
-kw = dict(n_agents=200, dur=5)
-
-
-def test_simple():
-    """Test the simple example via make_sim and hivsim.demo produce equivalent results."""
-    from hivsim_examples.simple.sim import make_sim
-
-    seed = 42
-    sim1 = make_sim(rand_seed=seed, **kw)
-    sim1.run()
-
-    sim2 = hivsim.demo('simple', run=False, rand_seed=seed, **kw)
-    sim2.run()
-
-    prev1 = sim1.results.hiv.prevalence[:]
-    prev2 = sim2.results.hiv.prevalence[:]
-    assert np.allclose(prev1, prev2), 'make_sim and hivsim.demo should produce identical results'
-
-
-def test_zimbabwe():
-    """Test the Zimbabwe example via make_sim and hivsim.demo produce equivalent results, and plot."""
-    from hivsim_examples.zimbabwe.sim import make_sim
-
-    seed = 42
-    sim1 = make_sim(rand_seed=seed, n_agents=500, stop=1995)
-    sim1.run()
-
-    sim2 = hivsim.demo('zimbabwe', run=False, rand_seed=seed, n_agents=500, stop=1995)
-    sim2.run()
-
-    prev1 = sim1.results.hiv.prevalence[:]
-    prev2 = sim2.results.hiv.prevalence[:]
-    assert np.allclose(prev1, prev2), 'make_sim and hivsim.demo should produce identical results'
-
-    # Check population is scaled to ~10M (Zimbabwe 1990)
-    n_alive_start = sim1.results.n_alive[0]
-    assert np.isclose(n_alive_start, 9_980_999, rtol=0.05), \
-        f'n_alive at t=0 ({n_alive_start}) should be close to 9,980,999'
-
-    # Run a full version (no plot when running via pytest)
-    sim3 = hivsim.demo('zimbabwe', plot=False)
-    return
-
-
-if __name__ == '__main__':
-    test_simple()
-    test_zimbabwe()
-    print('All tests passed!')
